Move alarm handler from scheduler to daemon.go
The scheduler is now only responsible for invoking a callback when an
alarm is triggered, improving the separation of conserns. The code in
daemon.go is the glue that connects the pieces and is better suited to
handle alarms. This fixes the issue of not invalidating the alarm cache
in the scheduler when a non-recurring alarm is removed from the DB.

Fixes #4